Add artist paginator to open studios catalog
Problem
--------

As a visitor to the open studios event I have clicked on one artist from the Catalog page. If there was a Next Artist button, I could thumb through all the participating artists without having to go back to the Catalog page and pick another artist.

Wouldn't that be fun?

Solution
--------

Build a paginator to figure out whos previous and next and then
render links to these artists on the artist/show page in the catalog.