Add last seen and since fields to admin artist csv download
As the event producer who would like to communicate with the artists who are planning on doing open studios, I'd like to also know who created a profile recently or updated their profile recently. Sometimes they sign up for the site and think that means they are in the show now.

Solution
--------

Add last seen and since and last updated profile to the csv download.