Fix GTM template
problem
--------

The GTM partial was causing some issues because the HTML tags where
prefixed with the slim `javascript:` tag which i think caused issues
with formatting. the page had the following

```
<noscript><iframe src='https://www.googletagmanager.com/ns.html?id=GTM-M7V3ZMB'
                  height= "0" width = "0" style = "display:none;visibility:hidden" > < /iframe></noscript >
```

which made the browser complain with
```
Uncaught SyntaxError: Unexpected token '<'
```

solution
---------

make the template a little dumber using erb so that all the tags in the
partial look like regular tags.