Fix missing favorite thumbs
Problem
----------

Some time ago, we dumped the `thumb` size for paperclip but we never
stopped using it.

Solution
---------

Move all cases where we used `thumb` to using `small` (mostly around
favorites) and while in there, use the `background_image_style` helper
more so that we have a consistent rendering of the inline background
image style which we use a bunch.