Remove Feedbacks controller and js in favor of NotifyMau component
Problem
--------

We had 2 basically identical ways to submit feedback.  Why not have just
1?

Solution
---------

Remove feedback specific controller and modal and form with the
NotifyMau component which is generalized so that it can support
different types of notification types.

This component writes to the same place as the previous feedback
controller did.  It just runs through a more generic FeedbackMail form.