Add tracking classes for gtm
problem
-------

we want to track folks with gtm to figure out where our users are
clicking on catalog pages...
mostly so we can report back to users : "hey, 10 people clicked on your
video" or "10 people clicked on your instagram link".

https://www.pivotaltracker.com/story/show/178997720

solution
--------

Add `gtm_` prefixed classes for links on the artist's catalog page.

The following classes can be used for tracking

* gtm_catalog_video_conference_url
* gtm_catalog_shop_url
* gtm_catalog_website_url
* gtm_catalog_instagram_url
* gtm_catalog_phone_link