Add flyout venmo qr code with link to venmo
Problem
---------

The donate link that hits straight to venmo doesn't really cover the
cases for tablets and phones and desktop.  For the first two you can
imagine that someone might want to scan the image with their phone.

Solution
----------

The button fly's out a qr code and if you click on the code it tries to
go to the app.  This should motivate folks to either scan the image if
they are on a desktop or tablet, and to just click if they're on their
phone.