Add notification banner model and view
Problem
-----------

When we need to do maintenance and other thing, we need to notify our users
(the ones that check in often) with a notification that stuffs going go happen.

we could email them but that's through MailChimp and it's likely that 70% of them don't care.

So let's focus on the common users (and visitors).

Solution
---------

After a discussion with @rdavid1099 , I decided on a simple-ish approach.

Add a `Notification` model which can be easily added with some console logic.  And if there is an active `Notification` show it on the app.

Some day in the future (if we care) we can add an admin interface.

Screenshots
-------------
Run this on the console
```
> Notification.create(message: "The site will be down for a bit.  Hold tight.  We'll b back soon", activated_at: Time.current)
```
then

<img width="955" alt="Screen Shot 2022-07-12 at 12 31 19 AM" src="https://user-images.githubusercontent.com/427380/178434502-60b0a2ff-9dde-405e-a8d4-b8fe5f1b5131.png">