Guard against active storage issues.
Problem
-------

The recent deploy is allowing uploads of images which successfully end
up in amazon but when requesting variants we get a Aws::S3::Errors::BadRequest.

Solution
----------

First cut will be to guard against that and fallback to the old image.
Also log when this happens loudly.

Then maybe we can figure out why variants are not working properly

Changes
--------

Wrap image fetchers with a begin/rescue