Compartmentalize image attachment logic
Problem
----------

Now that we've got the duplicate active storage and paperclip stuff
and we're going to try to unwind the paperclip,
i wanted to put all the attachment stuff together.

Solution
---------

Build a concern which User, Artist, Studio and ArtPiece can use
to compartmentalize the swapping between ActiveStorage attachments
and paperclip.

This way unwinding should be easier