Fix issue with favorite images and attachments
problem
--------

With the new attachment mixin, `MauFan`s (which are now deprecated)
do not know how to find their image.  This lead to breaking pages
where there was an fan favorite trying to render it's profile image.

solution
---------

Don't render a fan's profile picture and don't link to their page
because it has the same issues.