Unflake cucumber specs
Problem
----------

Faker puts apostrophe's in company names and last names.
Cucumber matchers get a little upset with that and can't find them on
the page even though screenshots show they are there.

Solution
----------

Just rip apostrophe's out of names in the factory.  we can test
apostrophe's if we care in unit tests.

This should unflake common flakes in cucumber.