Reinstate the map on the Open Studios promo page
Problem
----------

with Covid, we took out the map cuz virtual.

we're out of that and back in person so ... let's turn the map back on.

Solution
---------

Revamp the map javascript to be a little more modular and put the map
back on the open studios promo page.  Now instead of adding tabs, just
add the map above the list of participants.