Support individual image fetch
problem
---------

The site is slow because when we ask for images, we were running the
serializer for different models and that serializer was asking for *all*
active storage sizes which meant it had to process all iamges even if we
didn't need them all.

Solution
---------

Because we don't (usually) need all versions, we can instead surface a
new endpoint that just surfaces the image and size that we want at the
time.

This PR introduces that new backend controller and moves the frontend to
connect to there when it needs an image.

Additionally, this impacts the data that goes in to the search index.
So we updated the way search indexing happens by first only indexing the
image sizes used by search results and also moving the 'reindexing' to a
job because when you update art, the ActiveStorage may not be
immediately available.  This makes sure that reindexing happens a short
time after any model updates so we can be more certain that the indexer
will properly pick up any changes from ActiveStorage.

Changes
--------

* remove images from models and presenters and serializers
* add new controller method to get an image by size for a model
* update frontend components to instead of looking for `images[size]` -
  fetch just the one you need
* Moved typing from JSX.Element to ElementType
* Add active job configuration
* Add jobs to index Artist, ArtPiece, Studio