Add custom auth header for api
problem
-------

With our dev server, we are behind basic auth.  We still want an api
consumer key that will prevent folks from pulling from the api and that
key should not be connected to the Basic Auth setup.

solution
--------

Add a secondary `mau-api-authorization` header that we can use for this
case that will allow access to the api even through the basic auth using
a separate key.