Add custom validation/cleaning for instagram
problem
---------

From the story - from trish
> I noticed a lot of artists are just typing in their instagram handle in their profile page however we expect a working url. So when their instagram link is displayed on their open studios profile page it goes to a broken page.
I have been fixing it for this fall 2021 open studios but it would be nice if we could somehow accept all of these things:

solution
------

Add custom parsing for instagram link field.  If it doesn't look like
instagram.com but add a www.instagram.com.  If it has a leading @, just
strip that.