Add api to get studio ids
problem
-------

for better caching, we'd like to just get a list of artist ids.  Then
when things get added or removed, it's easier to not have to pull the
whole set but just deal with the new items.

solution
---------

Add a new endpoint to get artist ids by studio

`/api/v2/artists/ids?studio=<studio>`

we have no need for *all* so studio is a required param.

And the endpoint is built to be resilient around bad studio ids.  just
return empty array if we can't figure it out.