Fix config for webhook
problem
------

we forgot that when we deploy, the base `config.yml` gets pushed.

this record didn't have the correct url for the webhook.

solution
-----

fix the deployed config.

notes
----

i manually fixed this on the server so this change isn't required right now but
will get picked up with our next deploy