Better handling of studios without artists with art
problem
-----

just ran into this odd case where we have a studio (development) with 1 artist who has no art.
the display was goofy because we didn't have the right clauses in place.

solution
------

show the 'more artists' *if* the studio has artists_without_art - it was likely a cut paste issue
and not too critical since most studios have artists with art.