Remove unused parse_key method
problem
-----

just digging around a bit and noticed this `parse_key` method which is
no longer used.  we used to not store open studios events but just pass
around some string keys.  those keys needed to be parsed to dates which
is what this function was for.

but for a long time we've been using the db (OpenStudiosEvent) and the
wrapper service (OpenStudiosEventService) so we don't need to handle
those string only keys anymore.

Also while in there i noticed that `for_display(reverse:)` seemed like
a weird parameter name.  So renamed it to `month_first` which is what
it actually means.

solution
----

* remove `parse_key` off the OpenStudiosEventService
* update all `for_display` methods around os services/objects to use `month_first:` instead of `reverse:` as the formatting argument
* replace a few direct calls to `OpenStudiosEvent.current in favor of wrapping with the EventService