More typescript and js correctness
problem
-------

we have typescript but we weren't really linting or checking against it except maybe in our editors.

solution
-------

this pr hardens things but resolving Typescript warnings and also adds a `react-hooks` lint rule that exposed a bunch of places we weren't doing the right stuff wrt `useMemo/Callback/Effect` and dependency lists.