Bug - setting activation tasks via rake doesn't clear the cache
problem
------

with the introduction of #448, we added a rake task to set activation dates for
all events.  This plumbing did not use our `OpenStudiosEventService` which provides an update method to manage the cache clearing on update.

solution
-------

revamp the rake task to use the service so when the updates happen, the open studios event cache is cleared.  In addition, while doing this, for visibility, I updated the admin index page to show when one is active, show which one in words and by highlighting the table.

This also takes out the link to the site preferences page as after all this, the `SitePreferences#open_studios_active` link is kind of moot.  Follow up PR will remove it.

Changes
------

* update open_studios:set_activation_dates rake task to use OpenStudiosEventService.update
* update admin index for events to highlight the activated open studios if there is one
* update the `@open_studios_active` variable present in all requests via Application controller to return a presenter
* update the OpenStudiosEventPresenter to provide and `#active?` method to report if the current os is active or not

Screenshots
--------