Build in support to move to opensearch
problem
-------

Elasticsearch used to be free, but it's getting harder to manage without some licensing.
OpenSearch is a fork that is designed to be used in open source projects.

This PR builds in support with a feature flag for us to move to OpenSearch.
There are only a few important but subtle mods that have to happen.

strategy
-------

Below the `Search::indexer` and `Search::QueryRunner` we need to build service classes that know which service is live (ES or OS) and based on that make the right calls with the right structure.

Previous search connections were relatively well compartmentalized except for the use of `elasticsearch-rails` which adds a bunch of search stuff to the models.  With this PR, that moves into the `Search::SearchableObject` which basically maintains settings for different search objects based on the model name but does not put it *in* the AR model code.

changes
------

* extract helper search service functions from `Search::Indexer`
* build `Search::SearchService` which reads the feature flag and decides whether or not to use OpenSearch or ES
* copy `elasticsearch-extensions` gem and make it ready for opensearch
* add Search::OpenSearch::SearchService and Search::OpenSearchClient
* add Search::SearchableObject to contain search settings for different models and not attach them to the AR model code
* add FeatureFlags.use_open_search?
* update circle to use `opensearchproject/opensource` docker image and update urls for testing
* add opensearch info to readme

notes
-----

After this PR, we can remove the elasticsearch model code additions and the elasticsearch gem

Deployment will mean shutting down ES and a notification should be present because during this time, search will not work and there *may* be issues uploading/updating art though it appears with some testing that the search service is safe against ES not working