Move from CircleCI to Github Actions.
problem
----

CircleCI is one more service we kind of don't need now that Github Actions has a nice
workflow setup for us.

solution
------

Move to Github Actions.

Changes
------

This PR started with a template combining the LocalTake and BryantStreetStudios project github workflow templates, and then added vips, opensearch and a few other little tweaks to get the whole thing rolling.

Only run on PR's to avoid too much build cost.  We have ~30hrs of build time per month.  For this build which might be about 25min of compute time, it means 60 builds per month so it'd be nice not to overrun that.