Fix up use of spinner in a couple places
problem
-----

in an effort to move away from jQuery I looked into our use of spinners
and found a couple places where we weren't maybe doing exactly the right thing.

current spinner in JS is purely divs and CSS so no fancy javascript needed.

but... we were doing a bad check for null vs empty in a couple spots.

solution
-----

leverage `isEmpty` better so we can more reliably show spinners when we need to.

i think this also cleans up a little more of the typescript complaining.