Add custom useWhateverContext hooks for contexts/providers
problem
------

We use a couple contexts.  To do this we need to

1. import the context
1. import react's `useContext`
1. setup the context

As a better practice and also to avoid the amount of info a consumer needs about what it's consuming, if the context provided it's own hook to provide that context, the callers don't need to import `useContext` and don't need to know the context name.

solution
------

clean it up!

changes
------

* add useSearchResultsContext hook
* add useArtPiecesContext hook
* use them
* did a little other ts cleanup which i noticed while checking my typescript of the context moves