Fix mocks now that we're using open search *and* elasticsearch
problem
----

after running some tests locally, I noticed that the studio spec update was taking a long time.
It seems that we were calling open search and failing even when we thought we weren't.
it appears the 404 was causing the update to take a long time.

After digging around, i found that the elastic search mocking we had in place before wasn't actually working because we didn't update the way we mocked things.

solution
----

update the elasticsearch test helpers to be more ready for open search which is what we're using now(ish)