Remove oskey from frontend
problem
------

we only need the os key for a couple things.  OpenStudiosTally is using it as a foreign key
to track sign ups (sad face).  on the frontend we had a generator because previously we didn't actually store the date/time but now that we have a proper OpenStudiosEvent (from 10 years ago eb17bde27c), we can use the event id and date time.

solution
-------

add before_validate to set the key based on start date and move on.