Fix up featured on instagram/facebook admin snippets
problem
----
as an admin user
when I visit the admin/artists page and then go to see the artist details by clicking the 'info' icon
then I see a bunch of helper text for if I'm going to post to instagram or Facebook or twitter.
so that we can better promote our events.

fixes #444

solution
---
that content needs to be updated.

the url should be changed from http://MissionArtists.org/ to https://www.missionartists.org (lowercase and https and www).
the open studios dates should use the OpenStudiosEventService.current.

If there is no open studios, the sample should just include the link to the artist