Cache current os for 1 day
problem
-------

if the current open studios is cached and that cache lasts beyond the "active" time, there are places where we just say if we got a current thing, we show stuff. The current thing is cached and does not become uncached when the end of activation date passes.

instead if we make the cache expire every 24 hours, then at some point it'll auto clear and then we can make sure that within 1 day of the end activation we'll get the current version that will be "inactive" and current will return nil.

solution
----

Add a cache expiry so the value will expire every day

fixes #503