Fix admin edit page
problem
-----

the admin artist edit page was broken.

we were still using an `OpenStruct` so if the user had links, the page wouldn't render
because ostruct is no longer `required` by default.

solution
-----

replace the OpenStruct with a new helper model and update system specs to make
artists that have some links created by default so that the test will prove out these
failures in the future.